# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SowingCreate": {
        "additionalProperties": false,
        "properties": {
          "environment": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment"
          },
          "germinated_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Germinated Count"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/SowingLifecycle",
            "default": "active"
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "method_container": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method Container"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "pretreatment": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pretreatment"
          },
          "quantity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SowingQuantity-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "seed_lot_id": {
            "format": "uuid",
            "title": "Seed Lot Id",
            "type": "string"
          },
          "sowing_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "substrate": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Substrate"
          },
          "temperature_max_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Max C"
          },
          "temperature_min_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Min C"
          }
        },
        "required": [
          "seed_lot_id"
        ],
        "title": "SowingCreate",
        "type": "object"
      },
      "SowingLifecycle": {
        "enum": [
          "active",
          "completed",
          "failed",
          "abandoned"
        ],
        "title": "SowingLifecycle",
        "type": "string"
      },
      "SowingLocationSummary": {
        "properties": {
          "display_path": {
            "title": "Display Path",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_path"
        ],
        "title": "SowingLocationSummary",
        "type": "object"
      },
...
      "SowingResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "environment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment"
          },
          "germinated_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Germinated Count"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/SowingLifecycle"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SowingLocationSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "method_container": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method Container"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "pretreatment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pretreatment"
          },
          "quantity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SowingQuantity-Output"
              },
              {
                "type": "null"
              }
            ]
          },
          "seed_lot": {
            "$ref": "#/components/schemas/SowingSeedLotSummary"
          },
          "seed_lot_id": {
            "format": "uuid",
            "title": "Seed Lot Id",
            "type": "string"
          },
          "sowing_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "substrate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Substrate"
          },
          "temperature_max_c": {
            "anyOf": [
              {
                "description": "An exact decimal serialized as a JSON string.",
                "pattern": "^-?(0|[1-9][0-9]*)(\\.[0-9]+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Max C"
          },
          "temperature_min_c": {
            "anyOf": [
              {
                "description": "An exact decimal serialized as a JSON string.",
                "pattern": "^-?(0|[1-9][0-9]*)(\\.[0-9]+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Min C"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "seed_lot_id",
          "seed_lot",
          "label",
          "sowing_date",
          "quantity",
          "germinated_count",
          "location_id",
          "location",
          "substrate",
          "method_container",
          "pretreatment",
          "temperature_min_c",
          "temperature_max_c",
          "environment",
          "lifecycle",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "SowingResponse",
        "type": "object"
      },
...
      "SowingUpdate": {
        "additionalProperties": false,
        "properties": {
          "environment": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Environment"
          },
          "germinated_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Germinated Count"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "lifecycle": {
            "$ref": "#/components/schemas/SowingLifecycle",
            "default": "active"
          },
          "location_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Id"
          },
          "method_container": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method Container"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "pretreatment": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pretreatment"
          },
          "quantity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SowingQuantity-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "seed_lot_id": {
            "format": "uuid",
            "title": "Seed Lot Id",
            "type": "string"
          },
          "sowing_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDate"
              },
              {
                "type": "null"
              }
            ]
          },
          "substrate": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Substrate"
          },
          "temperature_max_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Max C"
          },
          "temperature_min_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature Min C"
          }
        },
        "required": [
          "seed_lot_id"
        ],
        "title": "SowingUpdate",
        "type": "object"
      },
...
      "SupplierKind": {
        "enum": [
          "seller",
          "nursery",
          "supermarket",
          "person",
          "exchange",
          "other"
        ],
        "title": "SupplierKind",
        "type": "string"
      },
      "SupplierResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/SupplierKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "retired_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retired At"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "website": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "website",
          "email",
          "phone",
          "notes",
          "retired_at",
          "created_at",
          "updated_at"
        ],
        "title": "SupplierResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Florabase API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sowings": {
      "get": {
        "operationId": "listSowings",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SowingResponse"
                  },
                  "title": "Response Listsowings",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List All",
        "tags": [
          "sowings"
        ]
      },
      "post": {
        "operationId": "createSowing",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SowingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SowingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "sowings"
        ]
      }
    },
    "/api/v1/sowings/{sowing_id}": {
      "get": {
        "operationId": "getSowing",
        "parameters": [
          {
            "in": "path",
            "name": "sowing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sowing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SowingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read",
        "tags": [
          "sowings"
        ]
      },
      "put": {
        "operationId": "updateSowing",
        "parameters": [
          {
            "in": "path",
            "name": "sowing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sowing Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SowingUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SowingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update",
        "tags": [
          "sowings"
        ]
      }
    },
    "/api/v1/suppliers": {
      "get": {
        "operationId": "listSuppliers",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SupplierResponse"
                  },
                  "title": "Response Listsuppliers",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List All",
        "tags": [
          "suppliers"
        ]
      },
      "post": {
        "operationId": "createSupplier",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "suppliers"
        ]
      }
    },
    "/api/v1/suppliers/{supplier_id}": {
      "get": {
        "operationId": "getSupplier",
        "parameters": [
          {
            "in": "path",
            "name": "supplier_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Supplier Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read",
        "tags": [
          "suppliers"
        ]
      },
      "put": {
        "operationId": "updateSupplier",
        "parameters": [
          {
            "in": "path",
            "name": "supplier_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Supplier Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update",
        "tags": [
          "suppliers"
        ]
      }
    },
    "/api/v1/suppliers/{supplier_id}/reactivate": {
      "post": {
        "operationId": "reactivateSupplier",
        "parameters": [
          {
            "in": "path",
            "name": "supplier_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Supplier Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reactivate",
        "tags": [
          "suppliers"
        ]
      }
    },
    "/api/v1/suppliers/{supplier_id}/retire": {
      "post": {
        "operationId": "retireSupplier",
        "parameters": [
          {
            "in": "path",
            "name": "supplier_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Supplier Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retire",
        "tags": [
          "suppliers"
        ]
      }
    }
  }
}